got cycling working again!

+++ src/process.h
@@ -40,6 +40,6 @@ uint8_t processSave();
 uint8_t processLoad();
 uint8_t processNew();
 uint8_t processHome();
-uint8_t processMove(uint16_t degree);
+uint8_t processMove(uint8_t degree);
 
 #endif
